requirements.txt
streamlit
langchain
langchain-community
langchain-groq
langgraph
pydantic
python-dotenv
chromadb
pypdf
